lazy load seach dialog
