+++ packages/docs/package.json
@@ -54,6 +54,7 @@
     "@vue/compiler-sfc": "^3.4.27",
     "@vuetify/api-generator": "workspace:*",
     "ajv": "^8.12.0",
+    "algoliasearch-helper": "^3.22.3",
     "async-es": "^3.2.5",
     "date-fns": "^3.6.0",
     "emailjs-com": "^3.2.0",
